Don't use set_unless
set_unless doesn't really do what you might think it does, and makes it
difficult to set the various attributes in a wrapper cookbook, so change
set_unless to a guard that checks if the attribute has been set instead.
Change is_local_host? to check for sockets in the form "host:/path/to/socket"
and return false (although I admit that symantic might be wrong)